refactor: 사진첩 권한 없는 경우 설정으로 이동하도록 수정
